File structure fixed. Still login issues.
