+++ pages/loginSequence/Name.js
@@ -1,44 +1,42 @@
 import React, { useState } from 'react';
 import { Text, TextInput, View } from 'react-native';
 import GestureRecognizer from 'react-native-swipe-gestures';
-import Styles from '../components/Styles';
-import InputStyles from '../components/InputStyles';
-import SwipeConfig from '../components/SwipeConfig';
+import Styles from '../../components/Styles';
+import SwipeConfig from '../../utils/SwipeConfig';
+import InputStyles from '../../components/InputStyles';
 
 /**
  * Greets the user in registration.
  * 
  * @param {Route} route
- * @param {Navigator} navigation 
+ * @param {Navigator} navigation
+ * @param {Object} userData
+ * @param {Callback} applyUserData
+ * @param {Callback} updateCurPage
  */
-const Handle = ({ route, navigation, userData, applyUserData, updateCurPage }) => {
+const Name = ({ route, navigation, userData, applyUserData, updateCurPage }) => {
   // Names are stateful
-  const [username, setUsername] = useState('');
-
-  const updateHandle = (text) => {
-    let usernameMinusAt = text;
-    if (text.charAt(0) == '@')
-      usernameMinusAt = usernameMinusAt.substr(1);
-    setUsername(usernameMinusAt);
-  };
+  const [firstName, setFirstName] = useState('');
+  const [lastName, setLastName] = useState('');
 
   // Proceed to the next step in registration.
   const proceed = () => {
     // Don't allow the user to continue if empty fields.
-    if (username == '')
+    if (firstName == '' || lastName == '')
       return;
 
     // Grab and update the data.
     const profileData = userData;
     
-    profileData.username = username;
+    profileData.firstName = firstName;
+    profileData.lastName = lastName;
 
     // Apply the new changes.
     applyUserData(profileData);
 
     // Update current page, and navigate.
     updateCurPage(route.params.pageIndex + 1);
-    navigation.navigate('Email', {
+    navigation.navigate('Username', {
       pageIndex: route.params.pageIndex + 1,
     });
   };
@@ -47,7 +45,7 @@ const Handle = ({ route, navigation, userData, applyUserData, updateCurPage }) =
   const backtrack = () => {
     // Update current page, and navigate.
     updateCurPage(route.params.pageIndex - 1);
-    navigation.navigate('Name', {
+    navigation.navigate('Greet', {
       pageIndex: route.params.pageIndex - 1,
     });
   };
@@ -60,21 +58,24 @@ const Handle = ({ route, navigation, userData, applyUserData, updateCurPage }) =
       style={Styles.container}
     >
       <Text style={[Styles.text, Styles.headerText]}>
-        What should we&nbsp;
+        What is your&nbsp;
         <Text style={Styles.blueAccentText}>
-        call&nbsp;
+        name?
         </Text>
-        you?
       </Text>
       <View style={{marginTop: 15}}/>
       <TextInput 
         style={InputStyles.textInput}
-        placeholder={'@handle'}
-        onChangeText={text => updateHandle(text)}
-        value={username !== '' ? '@' + username : null}
+        placeholder='john'
+        onChangeText={text => setFirstName(text)}
+      />
+      <TextInput 
+        style={InputStyles.textInput}
+        placeholder='smith'
+        onChangeText={text => setLastName(text)}
       />
     </GestureRecognizer>
   );
 };
  
-export default Handle;
+export default Name;
