+++ utils/SwipeConfig.js
